practice_from_books/pyproject.toml
[project]
name = "practice-from-books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pathlib>=1.0.1",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
